Reuse BasicFoodAutoComplete
